added ability to decode []string
